+++ typed_sql/lib/src/codegen/build_code.dart
@@ -314,7 +314,7 @@ Iterable<Spec> buildTable(ParsedTable table, ParsedSchema schema) sync* {
         'Unsupported type "${field.typeName}"',
       ),
     };
-    if (field.backingType == field.typeName) {
+    if (!field.isCustomType) {
       return readBackingType;
     }
     return '''
@@ -486,6 +486,90 @@ Iterable<Spec> buildTable(ParsedTable table, ParsedSchema schema) sync* {
               ],
             )
           '''),
+        ),
+      )
+      ..methods.add(
+        Method(
+          (b) => b
+            ..name = 'insertValue'
+            ..documentation('''
+              Insert row into the `${table.name}` table.
+
+              ${rowClass.fields.whereDefaultAndNullable.isEmpty ? '' : '''\n
+              > [!WARNING]
+              > It is not possible to insert the _default value_ for fields that
+              > are nullable. Providing `null` will insert `NULL` for
+              > ${rowClass.fields.whereDefaultAndNullable.map((f) => '`${f.name}`').join(', ')}.
+              '''}
+
+              Returns a [InsertSingle] statement on which `.execute` must be
+              called for the row to be inserted.
+            ''')
+            ..optionalParameters.addAll(
+              rowClass.fields.map((field) {
+                // Depending on whether a field has a default value and/or is
+                // nullable we have the following cases:
+                //
+                //  1. `!hasDefault && !isNullable`:
+                //     The user must give us a value!
+                //     The parameter is _required_ and non-nullable.
+                //
+                //  2. `hasDefault && !isNullable`:
+                //     The user may give us a value, or we can omit the field
+                //     and the database will insert the _default value_.
+                //     The parameter is optional and nullable, null means omit
+                //     the field when inserting the row.
+                //
+                //  3. `!hasDefault && isNullable`:
+                //     The user may give us a value, or we can insert `NULL` as
+                //     the default value -- this also what .insert() does!
+                //     The parameter is optional and nullable, null means set
+                //     the field `NULL` when inserting the row.
+                //
+                //  4. `hasDefault && isNullable`:
+                //     The user may give us a value, or omit the field to get
+                //     _default value_, or the user may give `null` meaning
+                //     `NULL`. We cannot represent all 3 options!
+                //     We always intepret `null` as `NULL`, thus, the user
+                //     cannot omit the field and get the _default value_.
+                //     The user can use `.insert()` instead of `.insertValue()`.
+                //     This could be surprising which is why we have:
+                //      * A warning in the documentation, and,
+                //      * Decided that the parameter is _required_ and nullable.
+                //     The parameter is _required_ and _nullable_, so that
+                //     inserting `null` is explicit, not implicit!
+                //
+                // NOTE: We do not set the default value as _default value_ in
+                //       dart, because this does not work for auto-increment or
+                //       for NOW() and similar annotations.
+                final isOptional = field.hasDefault ^ field.isNullable;
+                final nullablePostfix = field.hasDefault || field.isNullable
+                    ? '?'
+                    : '';
+
+                return Parameter(
+                  (b) => b
+                    ..name = field.name
+                    ..named = true
+                    ..required = !isOptional
+                    ..type = refer('${field.typeName}$nullablePostfix'),
+                );
+              }),
+            )
+            ..returns = refer('InsertSingle<$rowClassName>')
+            ..lambda = true
+            ..body = Code('''
+              \$ForGeneratedCode.insertInto(
+                table: this,
+                values: [${rowClass.fields.map((field) {
+              // Use .asExpr to support custom data types
+              // Use ?.asExpr if there is a default value and the field is not
+              // nullable.
+              final canOmit = field.hasDefault && !field.isNullable;
+              return '${field.name}${canOmit ? '?' : ''}.asExpr';
+            }).join(', ')},],
+              )
+            '''),
         ),
       )
       ..methods.add(
@@ -1265,6 +1349,21 @@ extension on ParsedRecord {
 
 extension on ParsedField {
   String get type => isNullable ? '$typeName?' : typeName;
+
+  bool get hasDefault => defaultValue != null || autoIncrement;
+
+  bool get isCustomType => backingType != typeName;
+}
+
+extension on List<ParsedField> {
+  /// Get a list of fields that both have a default value and are nullable.
+  ///
+  /// These fields are somewhat problematic to accept in `insertValue`, because
+  /// we map `null` to `NULL`, so there is no way to insert the and get the
+  /// default value. This kind of matters if the _default value_ is
+  /// auto-increment or `NOW()`. Hence, we should warn users.
+  List<ParsedField> get whereDefaultAndNullable =>
+      where((f) => f.hasDefault && f.isNullable).toList();
 }
 
 String backingExprType(String backingType) {
@@ -1284,7 +1383,7 @@ String backingExprType(String backingType) {
 
 String fieldType(ParsedField field) {
   var backingTypeName = backingExprType(field.backingType);
-  if (field.backingType == field.typeName) {
+  if (!field.isCustomType) {
     return backingTypeName;
   }
   return '${field.typeName}Ext._exprType';