+++ typed_sql/test/typed_sql/custom_types/custom_integer/custom_integer_test.g.dart
@@ -103,6 +103,16 @@ extension TableItemExt on Table<Item> {
     required Expr<MyCustomType> value,
   }) => $ForGeneratedCode.insertInto(table: this, values: [id, value]);
 
+  /// Insert row into the `items` table.
+  ///
+  /// Returns a [InsertSingle] statement on which `.execute` must be
+  /// called for the row to be inserted.
+  InsertSingle<Item> insertValue({int? id, required MyCustomType value}) =>
+      $ForGeneratedCode.insertInto(
+        table: this,
+        values: [id?.asExpr, value.asExpr],
+      );
+
   /// Delete a single row from the `items` table, specified by
   /// _primary key_.
   ///