+++ typed_sql/test/typed_sql/unique/unique_model/unique_model_test.g.dart
@@ -121,6 +121,19 @@ extension TableAccountExt on Table<Account> {
     values: [accountId, accountNumber, balance],
   );
 
+  /// Insert row into the `accounts` table.
+  ///
+  /// Returns a [InsertSingle] statement on which `.execute` must be
+  /// called for the row to be inserted.
+  InsertSingle<Account> insertValue({
+    int? accountId,
+    required String accountNumber,
+    double? balance,
+  }) => $ForGeneratedCode.insertInto(
+    table: this,
+    values: [accountId?.asExpr, accountNumber.asExpr, balance?.asExpr],
+  );
+
   /// Delete a single row from the `accounts` table, specified by
   /// _primary key_.
   ///